added get cloudwatch log group tags
