+++ go.mod
@@ -3,7 +3,7 @@ module awstaghelper
 go 1.14
 
 require (
-	github.com/aws/aws-sdk-go v1.33.15
+	github.com/aws/aws-sdk-go v1.33.17
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/spf13/cobra v1.0.0
 	github.com/spf13/viper v1.7.0
