+++ libs/commonLib/aws.go
@@ -1,0 +1,23 @@
+package commonLib
+
+import (
+	"fmt"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+)
+
+func GetSession(region string, profile string) *session.Session {
+	sess, err := session.NewSessionWithOptions(session.Options{
+		Profile: profile,
+		Config: aws.Config{
+			Region: aws.String(region),
+		},
+	})
+
+	if err != nil {
+		fmt.Println("Error", err)
+		return nil
+	}
+
+	return sess
+}
