--- modules/rdsHelper/setTags.go
@@ -1,40 +1,0 @@
-package rdsHelper
-
-import (
-	"fmt"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/service/rds"
-	"github.com/aws/aws-sdk-go/service/rds/rdsiface"
-)
-
-// TagRds tag rds instances. Take as input data from csv file. Where first column arn
-func TagRds(csvData [][]string, client rdsiface.RDSAPI) {
-	var tags []*rds.Tag
-	for r := 1; r < len(csvData); r++ {
-		for c := 1; c < len(csvData[0]); c++ {
-			tags = append(tags, &rds.Tag{
-				Key:   &csvData[0][c],
-				Value: &csvData[r][c],
-			})
-		}
-
-		input := &rds.AddTagsToResourceInput{
-			ResourceName: aws.String(csvData[r][0]),
-			Tags:         tags,
-		}
-
-		_, err := client.AddTagsToResource(input)
-		if err != nil {
-			if aerr, ok := err.(awserr.Error); ok {
-				switch aerr.Code() {
-				default:
-					fmt.Println(aerr.Error())
-				}
-			} else {
-				fmt.Println(err.Error())
-			}
-			return
-		}
-	}
-}
